requirements.txt
streamlit==1.50.0
spacy==3.8.7
pandas==2.3.2
altair==5.5.0
PyPDF2==3.0.1
python-docx==1.2.0